Moved DataAssembler and hooks in same file. Moved to utils folder
