+++ typescript/packages/group-tree-plot/src/utils/DataAssembler.ts
@@ -9,8 +9,9 @@ import type {
 } from "../types";
 import _ from "lodash";
 import { printTreeValue } from "../utils";
+import React from "react";
 
-export default class DataAssembler {
+export class DataAssembler {
     datedTrees: DatedTree[];
     edgeMetadataList: EdgeMetadata[];
     nodeMetadataList: NodeMetadata[];
@@ -146,4 +147,39 @@ function findTreeAndDateIndex(
 
     // No matching entry found
     return [-1, -1];
+}
+export function useDataAssembler(
+    datedTrees: DatedTree[],
+    edgeMetadataList: EdgeMetadata[],
+    nodeMetadataList: NodeMetadata[]
+): DataAssembler | null {
+    const dataAssembler = React.useMemo(() => {
+        if (datedTrees.length === 0) return null;
+
+        const assembler = new DataAssembler(
+            datedTrees,
+            edgeMetadataList,
+            nodeMetadataList
+        );
+
+        return assembler;
+    }, [datedTrees, edgeMetadataList, nodeMetadataList]);
+
+    return dataAssembler;
+}
+export function useDataAssemblerTree(assembler: DataAssembler) {
+    return assembler.getActiveTree();
+}
+export function useDataAssemblerPropertyValue(
+    assembler: DataAssembler,
+    data: NodeData | EdgeData,
+    property: string
+): number | null {
+    return assembler.getPropertyValue(data, property);
+}
+export function useDataAssemblerTooltip(
+    assembler: DataAssembler,
+    data: NodeData | EdgeData
+): string {
+    return assembler.getTooltip(data);
 }
